Prevent draft deletion on publish submission
This change has been introduced to prevent drafts created through the
submission process (submissions) from being deleted when they are
published. In implementing this I have tightened up what actions can
happen in various different states.

The 'not-submitted' and 'submitted' statuses are only possible on drafts
that were created during the submissions process. They will not have a
service ID until they are published. Because only submissions with the
'submitted' status can be published only they can have service IDs. When
this kind of draft is published it is not deleted so as to preserve the
submission.

The 'enabled', 'disabled' and 'published' statuses are inherited from
the service model. Drafts can only have these statuses if they are
created from an existing service. These drafts will always have a
service ID. When this kind of draft is published it is deleted.

+++ app/main/views/drafts.py
@@ -36,7 +36,8 @@ def copy_draft_service_from_existing_service(service_id):
     ).first_or_404()
 
     draft_service = DraftService.query.filter(
-        DraftService.service_id == service_id
+        DraftService.service_id == service_id,
+        DraftService.status.notin_(('not-submitted', 'submitted')),
     ).first()
     if draft_service:
         abort(400, "Draft already exists for service {}".format(service_id))
@@ -222,6 +223,11 @@ def publish_draft_service(draft_id):
         DraftService.id == draft_id
     ).first_or_404()
 
+    if draft.status == 'not-submitted':
+        abort(400, "Cannot publish a draft if it is not submitted: {}".format(draft.status))
+    if draft.status == 'submitted' and draft.service_id:
+        abort(400, "Cannot re-publish a submitted service")
+
     if draft.service_id:
         service = Service.query.filter(
             Service.service_id == draft.service_id
@@ -236,14 +242,19 @@ def publish_draft_service(draft_id):
                                audit_data={'draftId': draft_id})
 
     try:
-        db.session.delete(draft)
+        if draft.status == 'submitted':
+            draft.service_id = service_from_draft.service_id
+            db.session.add(draft)
+        else:
+            db.session.delete(draft)
         db.session.commit()
     except IntegrityError as e:
         db.session.rollback()
+        action = 'update' if draft.status == 'submitted' else 'delete'
         current_app.logger.warning(
-            'Failed to delete draft {} after publishing service {}: {}'.format(
-                draft_id, service_from_draft.service_id, e.message)
-        )
+            'Failed to {action} draft {draft_id} after publishing service {service_id}: {error}'.format(
+                extra=dict(
+                    action=action, draft_id=draft_id, service_id=service_from_draft.service_id, error=e.message)))
 
     index_service(service_from_draft)
